pyproject.toml
[project]
name = "eschool"
version = "0.1.0"
description = "Système de gestion scolaire complet avec Django"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "django-extensions>=3.2.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "pillow>=10.0.0",
    "reportlab>=4.0.0",
    "python-decouple>=3.8",
    "django-crispy-forms>=2.0",
    "crispy-tailwind>=0.5.0",
    "django-htmx>=1.17.0",
    "django-allauth>=0.57.0",
    "whitenoise>=6.6.0",
    "gunicorn>=21.2.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "django-debug-toolbar>=4.2.0",
    "factory-boy>=3.3.0",
    "faker>=20.0.0",
]
